+++ src/ui/src/graphql/queries/onUpdateDocument.js
@@ -44,6 +44,8 @@ export default gql`
       HITLSectionsSkipped
       HITLReviewOwner
       HITLReviewOwnerEmail
+      HITLReviewedBy
+      HITLReviewedByEmail
       HITLReviewHistory
       TraceId
     }